% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_implicit.R
\docType{data}
\name{ft_implicit}
\alias{ft_implicit}
\title{filmtrust implicit dataset}
\format{A sparse matrix  of class \code{dgCMatrix} with 1508 users and 2071
  items. A (i,j) th element is rating provided by user i for item j. A
  (i,j) th element is zero if there is no rating value. There are 35494
  ratings ( sparsity = 0.0114 ). Source:
  \href{http://www.librec.net/datasets/filmtrust.zip}{librec datasets}}
\usage{
ft_implicit
}
\description{
filmtrust implicit dataset
}
\keyword{datasets}

