% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_cols.R
\name{top_cols}
\alias{top_cols}
\title{top_cols}
\usage{
top_cols(mat, row, k = 5, ignore)
}
\arguments{
\item{mat}{(Sparse matrix of class 'dgCmatrix' or a integer/numeric matrix or
'big.matrix') Rating matrix.}

\item{row}{(positive integer) Row number in which top columns are to be
selected.}

\item{k}{(positive integer) Number of column numbers to be recommended. This
might not be strictly adhered to, see Details.}

\item{ignore}{(integer vector) Column numbers to be ignored.}
}
\description{
Find the column numbers corresponding the largest values in a
  particular row of a matrix
}
\details{
To find top-n recommendations of a ratings matrix given an item (or
  a user).  Although k recommendations are expected to be returned, the
  function might sometimes return more or less than k recommendations.

  \itemize{ \item Less: This happens when it is not possible to recommend k
  elements. For example, k is larger than the number of elements. \item More:
  This happens when a few elements have same rating. The function returns the
  index corresponding to all the elements which have the same rating. If
  ratings were \code{3,2,2,2,3}: \itemize{ \item k = 3: returns
  \code{1, 5, 2, 3, 4} \item k = 2: returns \code{1, 5} \item k = 1: returns
  \code{1, 5} }}
}
\examples{
\dontrun{
require("slimrec")
data(ft_small)
dir.create(file.path(tempdir(), "slim_test2"))
temp <- slim(ft_small
             , returnMat = TRUE
             , directory = file.path(tempdir(), "slim_test2")
             )
unlink(file.path(tempdir(), "slim_test2"), recursive = TRUE)
dim(temp[["ratingMat"]])
top_cols(temp[["ratingMat"]], row = 53)
top_cols(temp[["ratingMat"]], row = 53, ignore = c(5,6))
}

}

